Windows support: prevent the creation of directories for UDS endpoints
