+++ pkg/common/cli/umask_windows.go
@@ -1,7 +1,13 @@
+//go:build windows
+// +build windows
+
 package cli
+
+import "github.com/sirupsen/logrus"
 
 const umaskSupported = false
 
-func setUmask(umask int) int {
-	return umask
+// SetUmask does nothing on Windows
+func SetUmask(log logrus.FieldLogger) {
+	// Nothing to do in this platform
 }
